Camera files being used currently.
Webstreaming.py - to use camera as a real time webcam, useful for focus adjustments. Must be on the same network as the pi, type in address bar "192.168.8.122:8000" to access.

timevideo.py - to take a video (10s) in .h264 format having the filename match the time the video was taken.

timecapture.py - to take an image (3280x2464 'max resolution') in .jpg format, having the filename match the time the image was taken.