Structure of Webserver
This python script represents the most basic necessities to run a webserver from the pi. There is NO FUNCTIONALITY other than reading what is written within the "html = " section. do_POST is used to send information to the server, but was omitted in this as it was not necessary to simply successfully connect to the webpage when on the same wifi signal and using the pi's IP address and the given port (8080).